lovely new parent class for paddles
